+++ src/ft_convert/ft_ptoa_base.c
@@ -6,7 +6,7 @@
 /*   By: mleblanc <mleblanc@student.42quebec.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/16 23:27:35 by mleblanc          #+#    #+#             */
-/*   Updated: 2021/08/24 16:48:42 by mleblanc         ###   ########.fr       */
+/*   Updated: 2021/10/05 00:38:33 by mleblanc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ char	*ft_ptoa_base(void *ptr, const char *base)
 	size_t	base_len;
 
 	base_len = ft_strlen(base);
-	len = ft_unumlen((size_t)ptr, base_len);
+	len = ft_unumlen((unsigned long long)ptr, base_len);
 	ret = ft_calloc(len + 1, sizeof(char));
 	if (!ret)
 		return (NULL);
